added puuid types
